src/main/java/com/sda/iteventapp/form/util/ValidatorUtil.java
package com.sda.iteventapp.form.util;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

public class ValidatorUtil {
    public static boolean isCorrectEmail(String email) {
        //RFC 5322
        String regex = "(?im)^(?=.{1,64}@)(?:(\"[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*\"@)|((?:[0-9a-z](?:\\.(?!\\.)|[-!#\\$%&'\\*\\+\\/=\\?\\^`\\{\\}\\|~\\w])*)?[0-9a-z]@))(?=.{1,255}$)(?:(\\[(?:\\d{1,3}\\.){3}\\d{1,3}\\])|((?:(?=.{1,63}\\.)[0-9a-z][-\\w]*[0-9a-z]*\\.)+[a-z0-9][\\-a-z0-9]{0,22}[a-z0-9])|((?=.{1,63}$)[0-9a-z][-\\w]*))$";
        Pattern pattern = Pattern.compile(regex);
        Matcher matcher = pattern.matcher(email);
        return matcher.matches() && email.replaceAll("[\\w\\@\\.\\-\\+]", "").length() == 0;
    }

    public static boolean isCorrectPassword(String password, int min, int max) {
        return password.length() >= min && password.length() <= max;
    }

    public static boolean isUsernameCorrect(String username) {
        return username != null;
    }
}
